Move CargoProfile into cargo_pgrx::cargo
